Rename CPU feature header and source files for consistency.
