+++ CMakeLists.txt
@@ -600,8 +600,8 @@ if(WITH_OPTIM)
                 endif()
             endif()
         endif()
-        list(APPEND ZLIB_ARCH_HDRS ${ARCHDIR}/arm.h)
-        list(APPEND ZLIB_ARCH_SRCS ${ARCHDIR}/armfeature.c)
+        list(APPEND ZLIB_ARCH_HDRS ${ARCHDIR}/arm_features.h)
+        list(APPEND ZLIB_ARCH_SRCS ${ARCHDIR}/arm_features.c)
         if(WITH_ACLE AND NOT MSVC AND NOT "${ARCH}" MATCHES "armv[2-7]")
             check_acle_compiler_flag()
             if(HAVE_ACLE_FLAG)
@@ -645,8 +645,8 @@ if(WITH_OPTIM)
             check_power8_intrinsics()
         endif()
         if(HAVE_VMX OR HAVE_POWER8_INTRIN)
-            list(APPEND ZLIB_ARCH_HDRS ${ARCHDIR}/power.h)
-            list(APPEND ZLIB_ARCH_SRCS ${ARCHDIR}/power.c)
+            list(APPEND ZLIB_ARCH_HDRS ${ARCHDIR}/power_features.h)
+            list(APPEND ZLIB_ARCH_SRCS ${ARCHDIR}/power_features.c)
         endif()
         # VMX specific options and files
         if(WITH_ALTIVEC)
@@ -686,7 +686,8 @@ if(WITH_OPTIM)
     elseif(BASEARCH_S360_FOUND)
         if(WITH_CRC32_VX)
             add_definitions(-DS390_FEATURES)
-            list(APPEND ZLIB_ARCH_SRCS ${ARCHDIR}/s390.c)
+            list(APPEND ZLIB_ARCH_HDRS ${ARCHDIR}/s390_features.h)
+            list(APPEND ZLIB_ARCH_SRCS ${ARCHDIR}/s390_features.c)
         endif()
         if(WITH_DFLTCC_DEFLATE OR WITH_DFLTCC_INFLATE)
             list(APPEND ZLIB_ARCH_SRCS ${ARCHDIR}/dfltcc_common.c)
@@ -712,8 +713,8 @@ if(WITH_OPTIM)
         endif()
     elseif(BASEARCH_X86_FOUND)
         add_definitions(-DX86_FEATURES)
-        list(APPEND ZLIB_ARCH_HDRS ${ARCHDIR}/x86.h)
-        list(APPEND ZLIB_ARCH_SRCS ${ARCHDIR}/x86.c)
+        list(APPEND ZLIB_ARCH_HDRS ${ARCHDIR}/x86_features.h)
+        list(APPEND ZLIB_ARCH_SRCS ${ARCHDIR}/x86_features.c)
         if(MSVC)
             list(APPEND ZLIB_ARCH_HDRS fallback_builtins.h)
         endif()
