+++ arch/arm/arm_features.h
@@ -1,4 +1,4 @@
-/* arm.h -- check for ARM features.
+/* arm_features.h -- check for ARM features.
  * For conditions of distribution and use, see copyright notice in zlib.h
  */
