+++ arch/power/power_features.c
@@ -1,4 +1,4 @@
-/* POWER feature check
+/* power_features.c - POWER feature check
  * Copyright (C) 2020 Matheus Castanho <msc@linux.ibm.com>, IBM
  * Copyright (C) 2021 Mika T. Lindqvist <postmaster@raasu.org>
  * For conditions of distribution and use, see copyright notice in zlib.h
@@ -6,7 +6,7 @@
 
 #include <sys/auxv.h>
 #include "../../zutil.h"
-#include "power.h"
+#include "power_features.h"
 
 Z_INTERNAL int power_cpu_has_altivec = 0;
 Z_INTERNAL int power_cpu_has_arch_2_07 = 0;
