+++ arch/x86/crc32_fold_pclmulqdq.c
@@ -24,7 +24,7 @@
 #include <wmmintrin.h>
 #include <smmintrin.h> // _mm_extract_epi32
 
-#include "x86.h"
+#include "x86_features.h"
 
 #include "../../crc32_fold.h"
