+++ arch/x86/x86_features.h
@@ -1,10 +1,10 @@
-/* x86.h -- check for CPU features
+/* x86_features.h -- check for CPU features
 * Copyright (C) 2013 Intel Corporation Jim Kukunas
 * For conditions of distribution and use, see copyright notice in zlib.h
 */
 
-#ifndef CPU_H_
-#define CPU_H_
+#ifndef X86_FEATURES_H_
+#define X86_FEATURES_H_
 
 extern int x86_cpu_has_avx2;
 extern int x86_cpu_has_avx512;
